drivers : wifi : Update RPU patch to rev#b73f836aaf6
LMAC:
The deagg module needs a word offset address. So the buff passed to
deagg is divided by 3. Due to an error, divide by 3 was performed twice
on the buff address. This caused the over writing of the buff passed to
deagg, leading to corrupted packets.

UMAC:
CAL-3331 : Tx Packet injector mode changes

Signed-off-by: Divya Sai Keerthi Tiruvakadu <divya.tiruvakadu@nordicsemi.no>